added label remapper for saving trainId labels

+++ data/transform/labels.py
@@ -31,7 +31,7 @@ class RemapLabels:
             self.mapping[i] = mapping[i] if i in mapping else ignore_id
 
     def _trans(self, labels):
-        labels = self.mapping[labels].astype(labels.dtype)
+        labels = self.mapping[labels] # .astype(labels.dtype)
         if self.ignore_id != self.ignore_class:
             labels[labels == self.ignore_id] = self.ignore_class
         return labels
